+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 This update introduces minor changes and improvements
 
+### Changed
+ - Refactor Configuration class to use FileFormat for serialization and deserialization
+
 
 ## [1.1.8] - 2025-03-21