+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 This update introduces minor changes and improvements
 
-### Changed
+### Changed 
  - Refactor Configuration class to use FileFormat for serialization and deserialization