+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.10] - 2025-05-20
+
+This update introduces a minor bug fix
+
+### Fixed
+ - Update file permissions to 0777 for configuration file handling
+
+
 ## [1.1.9] - 2025-05-19
 
 This update introduces minor changes and improvements